extract base class
