+++ sdk/identity/azure-identity/azure/identity/_internal/__init__.py
@@ -2,5 +2,5 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
-from .confidential_client_credential import ConfidentialClientCredential
+from .msal_credentials import ConfidentialClientCredential
 from .msal_transport_adapter import MsalTransportResponse, MsalTransportAdapter
